First Edit

+++ src/main/java/com/ptit/managecertificate/dao/UserDAO.java
@@ -8,7 +8,7 @@ import com.ptit.managecertificate.entity.User;
 public interface UserDAO {
 	public void save(User user);
 	public void update(User user);
-	public User findById(int id);
+	public User findById(Long id);
 	public void delete(User user);
 	public List<User> findAll();
 	public User getUserByUserName(String userName);
